# backend/src/services/auditlogs/meradhan-audit-api.openapi.yaml
openapi: 3.0.3
info:
  title: Meradhan Audit Log API
  description: |
    Comprehensive audit logging system for Meradhan frontend application.

    Features:
    - **Optional User ID**: Works for both authenticated and guest users
    - **Required Session**: Session tracking is mandatory for all operations
    - **Automatic Tracking**: Auto-increments page count and session duration
    - **Full Pagination**: All GET endpoints support pagination
    - **Browser Detection**: Captures browser, device, OS, and IP information
  version: 1.0.0
  contact:
    name: API Support
    email: support@meradhan.com

servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.meradhan.com
    description: Production server

tags:
  - name: Page Tracking
    description: Track user page views and interactions
  - name: Login Logs
    description: Retrieve login/logout activity logs
  - name: Session Logs
    description: Retrieve session tracking logs with page views

paths:
  /api/auditlogs/meradhan/page-tracking/start:
    post:
      tags:
        - Page Tracking
      summary: Start page view tracking
      description: |
        Initiates tracking for a new page view. Session ID is required.
        User ID is optional for guest users.
      operationId: startPageTrackingMeradhan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageViewRequest"
            examples:
              authenticated:
                summary: Authenticated user
                value:
                  sessionId: "session_abc123def456"
                  userId: 1
                  pagePath: "/dashboard"
                  pageTitle: "Dashboard"
                  entryTime: "2025-11-18T10:00:00.000Z"
                  scrollDepth: 0
                  interactions: 0
                  referrer: "https://google.com"
              guest:
                summary: Guest user
                value:
                  sessionId: "session_xyz789ghi012"
                  pagePath: "/products"
                  pageTitle: "Products"
                  entryTime: "2025-11-18T10:00:00.000Z"
                  scrollDepth: 0
                  interactions: 0
      responses:
        "200":
          description: Page view tracking started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageViewStartResponse"
              example:
                statusCode: 200
                message: "Meradhan Page View Log Created"
                success: true
                responseData:
                  pageViewId: 123
        "400":
          description: Bad request - Session ID is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                statusCode: 400
                message: "Session ID is required"
                success: false

  /api/auditlogs/meradhan/page-tracking/end/{pageId}:
    post:
      tags:
        - Page Tracking
      summary: End page view tracking
      description: |
        Ends tracking for a page view and records exit time, duration, scroll depth, and interactions.
        Session ID is required.
      operationId: endPageTrackingMeradhan
      parameters:
        - name: pageId
          in: path
          required: true
          description: The ID of the page view to end
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageViewEndRequest"
            example:
              sessionId: "session_abc123def456"
              userId: 1
              pagePath: "/dashboard"
              pageTitle: "Dashboard"
              exitTime: "2025-11-18T10:05:00.000Z"
              duration: 300
              scrollDepth: 85
              interactions: 12
      responses:
        "200":
          description: Page view tracking ended successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                statusCode: 200
                message: "Meradhan Page View Log Ended"
                success: true
        "400":
          description: Bad request - Session ID is required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/auditlogs/meradhan/page-tracking/update/{pageId}:
    post:
      tags:
        - Page Tracking
      summary: Update page view tracking
      description: |
        Updates page view tracking data (e.g., scroll depth, interactions).
        Allows partial updates.
      operationId: updatePageTrackingMeradhan
      parameters:
        - name: pageId
          in: path
          required: true
          description: The ID of the page view to update
          schema:
            type: integer
          example: 123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PageViewUpdateRequest"
            example:
              scrollDepth: 90
              interactions: 15
      responses:
        "200":
          description: Page view tracking updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                statusCode: 200
                message: "Meradhan Page View Log Updated"
                success: true

  /api/auditlogs/meradhan/login-logs:
    get:
      tags:
        - Login Logs
      summary: Get login logs
      description: |
        Retrieves paginated login/logout activity logs for Meradhan.
        All parameters are optional for flexible filtering.
      operationId: getLoginLogsMeradhan
      parameters:
        - name: userId
          in: query
          description: Filter by user ID (optional)
          schema:
            type: integer
          example: 1
        - name: startDate
          in: query
          description: Filter from date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-11-01T00:00:00.000Z"
        - name: endDate
          in: query
          description: Filter to date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-11-18T23:59:59.999Z"
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: pageSize
          in: query
          description: Number of items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
      responses:
        "200":
          description: Login logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginLogsResponse"
              example:
                statusCode: 200
                message: "Meradhan Login Logs Retrieved"
                success: true
                responseData:
                  data:
                    - id: 1
                      userId: 1
                      name: "John Doe"
                      email: "john@example.com"
                      ipAddress: "192.168.1.1"
                      browserName: "Chrome"
                      deviceType: "desktop"
                      operatingSystem: "Windows"
                      sessionType: "login"
                      success: true
                      createdAt: "2025-11-18T10:00:00.000Z"
                    - id: 2
                      userId: null
                      name: null
                      email: "guest@example.com"
                      ipAddress: "192.168.1.2"
                      browserName: "Firefox"
                      deviceType: "mobile"
                      operatingSystem: "Android"
                      sessionType: "logout"
                      success: true
                      createdAt: "2025-11-18T09:30:00.000Z"
                  meta:
                    total: 100
                    page: 1
                    pageSize: 20
                    totalPages: 5
                    hasNextPage: true
                    hasPrevPage: false

  /api/auditlogs/meradhan/session-logs:
    get:
      tags:
        - Session Logs
      summary: Get session logs
      description: |
        Retrieves paginated session logs with associated page views.
        Supports filtering by user ID, session token, tracking token, and date range.
      operationId: getSessionLogsMeradhan
      parameters:
        - name: userId
          in: query
          description: Filter by user ID (optional)
          schema:
            type: integer
          example: 1
        - name: sessionToken
          in: query
          description: Filter by session token
          schema:
            type: string
          example: "session_abc123def456"
        - name: trackingToken
          in: query
          description: Filter by tracking token
          schema:
            type: string
          example: "track_xyz789"
        - name: startDate
          in: query
          description: Filter from date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-11-01T00:00:00.000Z"
        - name: endDate
          in: query
          description: Filter to date (ISO 8601 format)
          schema:
            type: string
            format: date-time
          example: "2025-11-18T23:59:59.999Z"
        - name: page
          in: query
          description: Page number (1-based)
          schema:
            type: integer
            minimum: 1
            default: 1
          example: 1
        - name: pageSize
          in: query
          description: Number of items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 20
      responses:
        "200":
          description: Session logs retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionLogsResponse"
              example:
                statusCode: 200
                message: "Meradhan Session Logs Retrieved"
                success: true
                responseData:
                  data:
                    - id: 1
                      userId: 1
                      sessionToken: "session_abc123def456"
                      trackingToken: "track_xyz789"
                      ipAddress: "192.168.1.1"
                      browserName: "Chrome"
                      deviceType: "desktop"
                      operatingSystem: "Windows"
                      startTime: "2025-11-18T10:00:00.000Z"
                      endTime: "2025-11-18T11:00:00.000Z"
                      duration: 3600
                      totalPages: 15
                      endReason: "logout"
                      createdAt: "2025-11-18T10:00:00.000Z"
                      updatedAt: "2025-11-18T11:00:00.000Z"
                      user:
                        firstName: "John"
                        middleName: null
                        lastName: "Doe"
                      pageViews:
                        - id: 1
                          userId: 1
                          sessionId: "session_abc123def456"
                          pagePath: "/dashboard"
                          pageTitle: "Dashboard"
                          entryTime: "2025-11-18T10:00:00.000Z"
                          exitTime: "2025-11-18T10:05:00.000Z"
                          duration: 300
                          scrollDepth: 85
                          interactions: 12
                          referrer: "https://google.com"
                          createdAt: "2025-11-18T10:00:00.000Z"
                          updatedAt: "2025-11-18T10:05:00.000Z"
                  meta:
                    total: 50
                    page: 1
                    pageSize: 20
                    totalPages: 3
                    hasNextPage: true
                    hasPrevPage: false

components:
  schemas:
    PageViewRequest:
      type: object
      required:
        - sessionId
        - pagePath
        - entryTime
      properties:
        sessionId:
          type: string
          description: Session token (required)
          example: "session_abc123def456"
        userId:
          type: integer
          description: User ID (optional - for authenticated users)
          nullable: true
          example: 1
        pagePath:
          type: string
          description: URL path of the page
          example: "/dashboard"
        pageTitle:
          type: string
          description: Title of the page
          example: "Dashboard"
        entryTime:
          type: string
          format: date-time
          description: Time when user entered the page
          example: "2025-11-18T10:00:00.000Z"
        scrollDepth:
          type: integer
          description: Initial scroll depth percentage (0-100)
          minimum: 0
          maximum: 100
          default: 0
          example: 0
        interactions:
          type: integer
          description: Initial interaction count
          minimum: 0
          default: 0
          example: 0
        referrer:
          type: string
          description: Referrer URL
          example: "https://google.com"

    PageViewEndRequest:
      type: object
      required:
        - sessionId
      properties:
        sessionId:
          type: string
          description: Session token (required)
          example: "session_abc123def456"
        userId:
          type: integer
          description: User ID (optional)
          nullable: true
          example: 1
        pagePath:
          type: string
          example: "/dashboard"
        pageTitle:
          type: string
          example: "Dashboard"
        exitTime:
          type: string
          format: date-time
          description: Time when user left the page
          example: "2025-11-18T10:05:00.000Z"
        duration:
          type: integer
          description: Time spent on page in seconds
          example: 300
        scrollDepth:
          type: integer
          description: Maximum scroll depth percentage (0-100)
          minimum: 0
          maximum: 100
          example: 85
        interactions:
          type: integer
          description: Total interaction count
          minimum: 0
          example: 12

    PageViewUpdateRequest:
      type: object
      properties:
        scrollDepth:
          type: integer
          description: Updated scroll depth percentage
          minimum: 0
          maximum: 100
          example: 90
        interactions:
          type: integer
          description: Updated interaction count
          minimum: 0
          example: 15
        duration:
          type: integer
          description: Updated duration in seconds
          example: 180

    PageViewStartResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Meradhan Page View Log Created"
        success:
          type: boolean
          example: true
        responseData:
          type: object
          properties:
            pageViewId:
              type: integer
              description: ID of the created page view
              example: 123

    SuccessResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Operation successful"
        success:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "Session ID is required"
        success:
          type: boolean
          example: false

    LoginLog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          nullable: true
          description: User ID (null for guest users)
          example: 1
        name:
          type: string
          nullable: true
          description: User name (null for guest users)
          example: "John Doe"
        email:
          type: string
          example: "john@example.com"
        ipAddress:
          type: string
          nullable: true
          example: "192.168.1.1"
        browserName:
          type: string
          nullable: true
          example: "Chrome"
        deviceType:
          type: string
          nullable: true
          enum: [desktop, mobile, tablet]
          example: "desktop"
        operatingSystem:
          type: string
          nullable: true
          example: "Windows"
        sessionType:
          type: string
          description: Type of session event
          enum: [login, logout, timeout, browser_close]
          example: "login"
        success:
          type: boolean
          description: Whether the login/logout was successful
          example: true
        createdAt:
          type: string
          format: date-time
          example: "2025-11-18T10:00:00.000Z"

    LoginLogsResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Meradhan Login Logs Retrieved"
        success:
          type: boolean
          example: true
        responseData:
          type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/LoginLog"
            meta:
              $ref: "#/components/schemas/PaginationMeta"

    SessionLog:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          nullable: true
          description: User ID (null for guest sessions)
          example: 1
        sessionToken:
          type: string
          description: Unique session identifier
          example: "session_abc123def456"
        trackingToken:
          type: string
          description: Tracking token for session
          example: "track_xyz789"
        ipAddress:
          type: string
          nullable: true
          example: "192.168.1.1"
        browserName:
          type: string
          nullable: true
          example: "Chrome"
        deviceType:
          type: string
          nullable: true
          enum: [desktop, mobile, tablet]
          example: "desktop"
        operatingSystem:
          type: string
          nullable: true
          example: "Windows"
        startTime:
          type: string
          format: date-time
          example: "2025-11-18T10:00:00.000Z"
        endTime:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-18T11:00:00.000Z"
        duration:
          type: integer
          description: Total session duration in seconds
          example: 3600
        totalPages:
          type: integer
          description: Total pages viewed in session
          example: 15
        endReason:
          type: string
          nullable: true
          description: Reason for session end
          enum: [logout, timeout, browser_close]
          example: "logout"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-18T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-18T11:00:00.000Z"
        user:
          type: object
          nullable: true
          description: User information (null for guest sessions)
          properties:
            firstName:
              type: string
              example: "John"
            middleName:
              type: string
              nullable: true
              example: null
            lastName:
              type: string
              example: "Doe"
        pageViews:
          type: array
          description: List of page views in this session
          items:
            $ref: "#/components/schemas/PageView"

    PageView:
      type: object
      properties:
        id:
          type: integer
          example: 1
        userId:
          type: integer
          nullable: true
          example: 1
        sessionId:
          type: string
          example: "session_abc123def456"
        pagePath:
          type: string
          example: "/dashboard"
        pageTitle:
          type: string
          nullable: true
          example: "Dashboard"
        entryTime:
          type: string
          format: date-time
          example: "2025-11-18T10:00:00.000Z"
        exitTime:
          type: string
          format: date-time
          nullable: true
          example: "2025-11-18T10:05:00.000Z"
        duration:
          type: integer
          description: Duration in seconds
          example: 300
        scrollDepth:
          type: integer
          description: Maximum scroll depth percentage
          example: 85
        interactions:
          type: integer
          description: Total interactions on the page
          example: 12
        referrer:
          type: string
          nullable: true
          example: "https://google.com"
        createdAt:
          type: string
          format: date-time
          example: "2025-11-18T10:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-11-18T10:05:00.000Z"

    SessionLogsResponse:
      type: object
      properties:
        statusCode:
          type: integer
          example: 200
        message:
          type: string
          example: "Meradhan Session Logs Retrieved"
        success:
          type: boolean
          example: true
        responseData:
          type: object
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/SessionLog"
            meta:
              $ref: "#/components/schemas/PaginationMeta"

    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
          example: 100
        page:
          type: integer
          description: Current page number
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          example: 20
        totalPages:
          type: integer
          description: Total number of pages
          example: 5
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPrevPage:
          type: boolean
          description: Whether there is a previous page
          example: false

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication (optional for guest users)

security:
  - bearerAuth: []